+++ src/api/objects/toot.ts
@@ -36,13 +36,13 @@ import {
     DEFAULT_FONT_SIZE,
     MEDIA_TYPES,
     VIDEO_TYPES,
-    arrowed,
     at,
     collapseWhitespace,
     determineMediaCategory,
     extractDomain,
     htmlToParagraphs,
     htmlToText,
+    optionalSuffix,
     removeDiacritics,
     removeEmojis,
     removeLinks,
@@ -667,7 +667,7 @@ export default class Toot implements TootObj {
         } else if (tagType == TypeFilterName.TRENDING_TAGS) {
             tags = this.trendingTags || [];
         } else {
-            tootLogger.warn(`${arrowed('containsTagsOfTypeMsg()')} called with invalid tagType: ${tagType}`);
+            tootLogger.warn(`containsTagsOfTypeMsg() called with invalid tagType: ${tagType}`);
             return;
         }
 
@@ -696,7 +696,7 @@ export default class Toot implements TootObj {
     private contentWithCard(): string {
         if (!this.contentCache[TootCacheKey.CONTENT_WITH_CARD]) {
             const cardContent = [this.card?.title || "", this.card?.description || ""].join(" ").trim();
-            const txt = (this.contentString() + (cardContent.length ? ` (${htmlToText(cardContent)})` : "")).trim();
+            const txt = (this.contentString() + optionalSuffix(cardContent, htmlToText)).trim();
             this.contentCache[TootCacheKey.CONTENT_WITH_CARD] = removeDiacritics(txt);
         }