+++ src/helpers/logger.ts
@@ -4,9 +4,10 @@
 import { ageString } from './time_helpers';
 import { isDebugMode, isDeepDebug } from './environment_helpers';
 import { sortKeysByValue, split } from './collection_helpers';
-import { TELEMETRY, arrowed, bracketed, createRandomString, isEmptyStr } from './string_helpers';
+import { TELEMETRY, arrowed, bracketed, createRandomString, isEmptyStr, optionalSuffix } from './string_helpers';
 import { type StringNumberDict } from '../types';
 
+type ErrorArgs = {args: any[], error?: Error};
 type LoggerArg = string | boolean | null | undefined;  // boolean so we can filter out optional args that are falsey
 
 const PREFIXERS = [
@@ -20,7 +21,7 @@ const PREFIXERS = [
     (str: string) => `~${str}~`,
     (str: string) => `-${str}-`,
     (str: string) => `#${str}#`,
-];
+] as const;
 
 /**
  * Standardized logger for consistent, prefixed, and optionally colorized logging throughout the application.
@@ -32,7 +33,7 @@ const PREFIXERS = [
  */
 export class Logger {
     logPrefix: string;
-    prefixes: string[];
+    private prefixes: string[];
 
     /**
      * Constructs a Logger instance with the given name and optional additional prefixes.
@@ -50,6 +51,7 @@ export class Logger {
 
     /**
      * Alternate constructor; makes the first two arguments into a parenthesized bracketed string.
+     * e.g. the prefix will be [name (parenthesized)].
      * @param {string} name - The main name for the logger.
      * @param {string} parenthesized - The value to parenthesize in the prefix.
      * @param {...string} args - Additional prefix arguments.
@@ -61,30 +63,24 @@ export class Logger {
 
     /**
      * Logs an error message or Error object to the console with the logger's prefix.
-     * Checks whether the 2nd arg is an instance of Error for special handling.
+     * Checks whether any element of 'args' is an instance of Error for special handling.
      * @param {string|Error} msg - The error message or Error object.
      * @param {...any} args - Additional arguments to log.
      * @returns {string} The error message string.
      */
     error(msg: string | Error, ...args: any[]): string {
-        if (msg instanceof Error) {
-            console.error(this.line(msg.message), ...args);
-            return msg.message;
-        }
-
-        msg = this.errorStr(msg, ...args);
-        console.error(this.line(msg), ...args);
+        const allArgs = [msg, ...args];
+        msg = this.errorStr(...allArgs);
+        console.error(this.line(msg), ...allArgs);
         return msg;
     }
 
     /**
-     * Logs a warning message to the console with the logger's prefix.
-     * Checks the 2nd arg in the same way as `error()`.
+     * Call console.warn() with the logger's prefix. Checks for Error objs in args in the same way as `error()`.
      * @param {string} msg - The warning message.
      * @param {...any} args - Additional arguments to log.
      */
-    warn =  (msg: string, ...args: any[]) => console.warn(this.line(this.errorStr(msg, ...args)), ...args);
-
+    warn =  (msg: string, ...args: any[]) => console.warn(this.line(this.errorStr(...[msg, ...args])));
     /** console.log() with the logger's prefix. */
     log =   (msg: string, ...args: any[]) => console.log(this.line(msg), ...args);
     /** console.info() with the logger's prefix. */
@@ -111,24 +107,18 @@ export class Logger {
      * Logs an error message and throws an Error with the stringified arguments and message.
      * @param {string} msg - The error message.
      * @param {...any} args - Additional arguments to include in the error.
-     * @throws {Error} Throws an error with the formatted message.
+     * @throws {Error} A new Error with the formatted message, optionally including the first Error argument.
      */
     logAndThrowError(msg: string, ...args: any[]): never {
         console.error(msg, args);
+        const errorArgs = this.findErrorArg(args);
 
-        if (args.length > 0) {
-            const [errorArgs, otherArgs] = split(args, arg => arg instanceof Error);
-
-            if (errorArgs.length > 0) {
-                msg = this.makeErrorMsg(errorArgs[0], msg);
-            }
-
-            if (otherArgs.length > 0) {
-                msg += [`, additional args:`, ...args.map(arg => JSON.stringify(arg, null, 4))].join(`\n`);
-            }
+        if (errorArgs.args.length > 0) {
+            msg += [`, additional args:`, ...args.map(arg => JSON.stringify(arg, null, 4))].join(`\n`);
         }
 
-        throw new Error(this.line(msg));
+        msg = this.line(msg);
+        throw errorArgs.error ? new Error(msg, {cause: errorArgs.error}) : new Error(msg);
     }
 
     /**
@@ -141,7 +131,7 @@ export class Logger {
     logArrayReduction<T>(before: T[], after: T[], objType: string, reason?: string): void {
         const numRemoved = before.length - after.length;
         if (numRemoved == 0) return;
-        this.trace(`Removed ${numRemoved} ${ reason ? (reason + " ") : ""}${objType}s leaving ${after.length}`);
+        this.trace(`Removed ${numRemoved}${optionalSuffix(reason)} ${objType}s leaving ${after.length}`);
     }
 
     /**
@@ -196,11 +186,36 @@ export class Logger {
      * @param {...any} args - Additional arguments.
      * @returns {string} The formatted error message.
      */
-    private errorStr(msg: string, ...args: any[]): string {
-        if (args[0] instanceof Error) {
-            return this.makeErrorMsg(args.shift() as Error, msg);
+    private errorStr(...args: any[]): string {
+        const errorArgs = this.findErrorArg(args);
+        const stringArgs = errorArgs.args.map(arg => typeof arg === 'string' ? arg : JSON.stringify(arg, null, 4));
+        const stringArg = stringArgs.length > 0 ? stringArgs.join(', ') : undefined;
+
+        if (errorArgs.error) {
+            return this.makeErrorMsg(errorArgs.error, stringArg);
         } else {
-            return msg;
+            if (!stringArg) this.warn(`errorStr() called with no string or error args, returning empty string`);
+            return stringArg || '';
+        }
+    }
+
+    /**
+     * Separate the Error type args from the rest of the args.
+     * @private
+     * @param {...any} args - Additional arguments.
+     * @returns {ErrorArgs} Object with `args` containing non-Error args and `error` if an Error was found.
+     */
+    private findErrorArg(args: any[]): ErrorArgs {
+        const [errorArgs, otherArgs] = split(args, arg => arg instanceof Error);
+
+        if (errorArgs.length > 0) {
+            if (errorArgs.length > 1) {
+                this.warn(`findErrorArg() called with multiple Error args, only using the first one:`, errorArgs);
+            }
+
+            return {args: otherArgs, error: errorArgs[0] as Error}
+        } else {
+            return {args: otherArgs};
         }
     }